change getDram return type

+++ scripts/main.d.ts
@@ -16,7 +16,7 @@ declare module 'mupen64plus-web' {
     hardReset(): void
     advanceFrame(): void
     reloadRom(romData: ArrayBuffer): Promise<void>
-    getDram(): Uint32Array
+    getDram(): Uint8Array
   }
 
   export interface CoreConfig {
